Refactor: 패키지 이동 및 회원가입 DTO 수정
- 패키지 삭제 및 이동
- SignupRequest의 role 프로퍼티를 Enum으로 받게끔 변경
