+++ src/main/kotlin/com/todo/todoapplication/domain/user/dto/request/SignupRequest.kt
@@ -1,7 +1,9 @@
 package com.todo.todoapplication.domain.user.dto.request
 
+import com.todo.todoapplication.domain.user.model.UserRole
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Pattern
 
 data class SignupRequest(
@@ -11,11 +13,11 @@ data class SignupRequest(
 
     @field:NotBlank
     @field:Pattern(
-        regexp="^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[!@#\$%^&*])[a-zA-Z0-9!@#\$%^&*]{8,20}\$",
+        regexp = "^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[@#$%^&+=!])(?=\\S+$).{8,15}$",
         message = "비밀번호는 8~20 소/대/특수문자를 혼합하여 작성해주세요"
     )
     val password: String,
 
-    @field:NotBlank
-    val role: String
+    @field:NotNull(message = "역할은 필수값입니다.")
+    val role: UserRole
 )
