+++ src/main/kotlin/com/todo/todoapplication/global/auth/CustomAccessDeniedHandler.kt
@@ -1,4 +1,4 @@
-package com.todo.todoapplication.global.auth.handler
+package com.todo.todoapplication.global.auth
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.todo.todoapplication.global.exception.response.ErrorResponse
@@ -6,23 +6,23 @@ import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
-import org.springframework.security.core.AuthenticationException
-import org.springframework.security.web.AuthenticationEntryPoint
+import org.springframework.security.access.AccessDeniedException
+import org.springframework.security.web.access.AccessDeniedHandler
 import org.springframework.stereotype.Component
 
 @Component
-class CustomAuthenticationEntryPoint : AuthenticationEntryPoint {
-    override fun commence(
+class CustomAccessDeniedHandler : AccessDeniedHandler {
+    override fun handle(
         request: HttpServletRequest,
         response: HttpServletResponse,
-        authException: AuthenticationException
+        accessDeniedException: AccessDeniedException
     ) {
-        response.status = HttpServletResponse.SC_UNAUTHORIZED
+        response.status = HttpServletResponse.SC_FORBIDDEN
         response.contentType = MediaType.APPLICATION_JSON_VALUE
         response.characterEncoding = "UTF-8"
 
         val objectMapper = ObjectMapper()
-        val jsonString = objectMapper.writeValueAsString(ErrorResponse.of(HttpStatus.UNAUTHORIZED, "JWT 확인 실패"))
+        val jsonString = objectMapper.writeValueAsString(ErrorResponse.of(HttpStatus.FORBIDDEN, "No permission to run API"))
         response.writer.write(jsonString)
     }
 }
