+++ src/main/kotlin/com/todo/todoapplication/global/config/ConfigurationPropertiesConfig.kt
@@ -1,6 +1,6 @@
 package com.todo.todoapplication.global.config
 
-import com.todo.todoapplication.global.auth.jwt.JwtProperties
+import com.todo.todoapplication.global.auth.JwtProperties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration
