+++ src/main/kotlin/com/todo/todoapplication/global/exception/auth/AccessDeniedApiException.kt
@@ -1,3 +1,3 @@
-package com.todo.todoapplication.global.exception
+package com.todo.todoapplication.global.exception.auth
 
 class AccessDeniedApiException(message: String) : RuntimeException(message)
