+++ src/main/kotlin/com/todo/todoapplication/global/exception/common/NoSuchEntityException.kt
@@ -1,4 +1,4 @@
-package com.todo.todoapplication.global.exception
+package com.todo.todoapplication.global.exception.common
 
 class NoSuchEntityException(
     message: String
