+++ src/main/kotlin/com/todo/todoapplication/global/exception/member/DuplicatedEmailException.kt
@@ -1,3 +1,3 @@
-package com.todo.todoapplication.global.exception
+package com.todo.todoapplication.global.exception.member
 
 class DuplicatedEmailException(message: String) : RuntimeException(message)
