+++ src/test/kotlin/com/todo/todoapplication/domain/todo/service/TodoServiceTest.kt
@@ -10,7 +10,7 @@ import com.todo.todoapplication.fixture.TodoFixture.Companion.todoResponse
 import com.todo.todoapplication.fixture.TodoFixture.Companion.todoResponseList
 import com.todo.todoapplication.fixture.TodoFixture.Companion.updateTodoRequest
 import com.todo.todoapplication.fixture.TodoFixture.Companion.wrongTodoId
-import com.todo.todoapplication.global.exception.NoSuchEntityException
+import com.todo.todoapplication.global.exception.common.NoSuchEntityException
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
